requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
requests==2.31.0
python-dotenv==1.0.1
lxml==5.1.0
regex==2023.12.25
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1